Add fuction move table

+++ DATN/DATN/Shared/RewardPointe.cs
@@ -12,7 +12,7 @@ namespace DATN.Shared
         [Key]
         public int RewardPointId { get; set; }
         public int CustomerId { get; set; }
-        public int RewardPoint { get; set; }
+        public decimal RewardPoint { get; set; }
         public DateTime UpdateDate { get; set; }
         public bool IsDeleted { get; set; }
         public int OrderId { get; set; }
